+++ src/Modules/Store/PickupValidationException.php
@@ -8,6 +8,7 @@ class PickupValidationException extends \InvalidArgumentException
     public const DUPLICATE_PICKUP_DAY_TIME = 'Multiply pickups for the same day and time.';
     public const PICK_UP_DATE_IN_THE_PAST = 'Pickup is in the past.';
     public const SLOT_COUNT_OUT_OF_RANGE = 'Slot value out of range.';
+    public const DESCRIPTION_OVERSIZED = 'Description is too long.';
     public const MORE_OCCUPIED_SLOTS = 'Slot value is smaller then occuied slots.';
     public const INVALID_STORE = 'Store does not exist.';